refactor(rust): cycle 42 LOC≤250 — split startup_routes.rs into submodules
Split 292-LOC startup_routes.rs into 6 submodules under startup_routes/:
- docs.rs (14), auth.rs (34), diag.rs (52), mailbox.rs (60),
  admin.rs (73), external.rs (76), mod.rs (17).

Public API preserved via pub(crate) re-exports. No behavior change.
